Moved ProfileFilter into separate file - filter_profile.go
